Disable c10::import(ns)
This disables namespaced c10::import which is broken with custom
mobile op builds.  This is to help prevent people from accidentally
breaking the custom mobile build in a mysterious way; if they use
the longform version it will work.  Fixing the analyzer is tracked
in https://github.com/pytorch/pytorch/issues/35397

Signed-off-by: Edward Z. Yang <ezyang@fb.com>

ghstack-source-id: 3a29e26a01484ecf59e341edd959cf924d346b56
Pull Request resolved: https://github.com/pytorch/pytorch/pull/35398